Refactor validation to decorators
